requirements.txt
# Core Framework
anthropic>=0.18.0
openai>=1.12.0
slack-sdk>=3.26.0
slack-bolt>=1.18.0

# Database
sqlalchemy>=2.0.25

# Data Validation
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
python-dateutil>=2.8.2
pytz>=2024.1

# Scheduling
apscheduler>=3.10.4

# Logging
loguru>=0.7.2
